arch(#706 v10): relocate checkProtocolRange to protocol root barrel + extended hand-authored docs sweep
v10 revisions to address codex r9 round-9 (1 P2 + 1 P3):

- Export `checkProtocolRange` for testability (codex r9 P2 #1):
  v9 parameterized the function over `serverVersion` (signature
  testable in principle) but kept it private in `connect.handlers.ts`
  so tests couldn't import without an illegal seam. v10 relocates
  `checkProtocolRange` + `failProtocolMismatch` to
  `packages/protocol/src/version.ts` and re-exports
  `checkProtocolRange` from the protocol root barrel.

  - `version.ts`: imports `ProtocolMismatchError` + `ProtocolMismatchReason`
    from `./network/methods.js`; defines `checkProtocolRange` (exported)
    + `failProtocolMismatch` (module-private; only called by
    `checkProtocolRange`).
  - `protocol/src/index.ts`: root barrel re-exports
    `checkProtocolRange` alongside `PROTOCOL_VERSION` +
    `compareProtocolVersion`.
  - `connect.handlers.ts`: drops local definitions, imports
    `checkProtocolRange` from `@moltzap/protocol`. The
    `yield* checkProtocolRange(params, PROTOCOL_VERSION)` call site
    is unchanged in shape.

  Tests can now `import { checkProtocolRange } from "@moltzap/protocol"`
  and inject future-version values directly:
    Effect.runSync(Effect.either(checkProtocolRange(
      { minProtocol: "2026.526.0", maxProtocol: "2026.526.0" },
      "2026.527.0",
    )))
    → result._tag === "Left"
    → result.left._tag === "ProtocolMismatchError"
    → result.left.data?.reason === "server-above-client-max"

- Extended hand-authored docs sweep (codex r9 P3 #1):
  v9 stub-edited docs/concepts/presence.mdx + docs/guides/presence.mdx
  but missed docs/architecture.mdx, docs/introduction.mdx, and the
  conformance comments/README mentioning PresenceUpdate / setOnline.

  v10 stub-edits each remaining stale reference:
  - docs/architecture.mdx: PresenceService row updated to name the
    post-v10 subscriber-registry role; impl-staff (#712) replaces
    with PresenceProjection row.
  - docs/introduction.mdx: "Online/offline/away" → "Online/working/
    offline (server-derived from lease lifecycle)".
  - conformance README/MODULE.md/index.ts: PresenceUpdate mentions
    updated to name `presence/changed` server-emitted notifications;
    flagged `setOnline` ref in `presence-same-state-no-double-fire.ts`
    as impl-staff scope.
  - presence-same-state-no-double-fire.ts: JSDoc rewritten to name
    `presenceProjection.onAgentConnect` (architect plan v6+);
    registerProperty label updated from "redundant setOnline" to
    "redundant onAgentConnect".

  Generated docs (docs/modules/protocol/src.mdx,
  docs/modules/protocol/testing/conformance/network.mdx,
  packages/protocol/src/MODULE.md,
  packages/protocol/src/testing/conformance/network/MODULE.md)
  auto-regenerated by pre-commit `pnpm docs:generate`.

Empirically verified:
- pnpm -r exec tsc --build --force: EXIT=0
- pnpm -r exec eslint src --max-warnings=0: EXIT=0
- `checkProtocolRange` importable from `@moltzap/protocol` root
  barrel; full repo `tsc --build` clean across all packages.

Design doc v10: posted as a comment on parent epic #710.

Architect sub-issue: #711.
