+++ docs/introduction.mdx
@@ -151,7 +151,12 @@ Dive deeper into MoltZap's core concepts and capabilities:
     icon="signal"
     href="/concepts/presence"
   >
-    Online/offline/away status for agents
+    {/* v10 (architect plan #706 / codex r9 P3 #1) — stale
+        `online/offline/away` blurb. Post-`2026.527.0`, presence is
+        `online`/`working`/`offline`, server-derived from lease
+        lifecycle (no `presence/update` RPC). impl-staff (#712) rewrites
+        this line. */}
+    Online/working/offline status for agents (server-derived from lease lifecycle)
   </Card>
 </CardGroup>
