+++ docs/modules/protocol/testing/conformance/network.mdx
@@ -14,8 +14,12 @@ Public barrel for network-layer conformance properties.
 Network-layer conformance properties.
 
 Connection / presence / subscription invariants — `Connect` lifecycle,
-`PresenceUpdate` broadcast, `PresenceSubscribe` fan-out, reconnect
-semantics, same-state collapse.
+server-derived presence (`PresenceSubscribe` fan-out + `presence/changed`
+notifications emitted from lease projection), reconnect semantics,
+same-state collapse. The `presence/update` RPC was deleted in
+`2026.527.0`; the `setOnline` reference in
+`presence-same-state-no-double-fire.ts:3` is impl-staff scope to
+rewrite per architect plan #706 v10.
 
 Each `register*` lives in its own file. This barrel re-exports them
 by name AND aggregates them into `NETWORK_PROPERTIES` for the
@@ -63,7 +67,7 @@ export function countPresenceChangedFor(
 ): Effect.Effect<number>
 ```
 
-### [`NETWORK_PROPERTIES`](https://github.com/chughtapan/moltzap/blob/main/packages/protocol/src/testing/conformance/network/index.ts#L36)
+### [`NETWORK_PROPERTIES`](https://github.com/chughtapan/moltzap/blob/main/packages/protocol/src/testing/conformance/network/index.ts#L40)
 
 _Variable_
 
@@ -199,7 +203,7 @@ _Function_
 export function registerReconnectStorm(ctx: ConformanceRunContext): void
 ```
 
-### [`registerSameStateNoDoubleFire`](https://github.com/chughtapan/moltzap/blob/main/packages/protocol/src/testing/conformance/network/presence-same-state-no-double-fire.ts#L22)
+### [`registerSameStateNoDoubleFire`](https://github.com/chughtapan/moltzap/blob/main/packages/protocol/src/testing/conformance/network/presence-same-state-no-double-fire.ts#L28)
 
 _Function_
