+++ packages/protocol/src/testing/conformance/network/index.ts
@@ -4,8 +4,12 @@
  * Network-layer conformance properties.
  *
  * Connection / presence / subscription invariants — `Connect` lifecycle,
- * `PresenceUpdate` broadcast, `PresenceSubscribe` fan-out, reconnect
- * semantics, same-state collapse.
+ * server-derived presence (`PresenceSubscribe` fan-out + `presence/changed`
+ * notifications emitted from lease projection), reconnect semantics,
+ * same-state collapse. The `presence/update` RPC was deleted in
+ * `2026.527.0`; the `setOnline` reference in
+ * `presence-same-state-no-double-fire.ts:3` is impl-staff scope to
+ * rewrite per architect plan #706 v10.
  *
  * Each `register*` lives in its own file. This barrel re-exports them
  * by name AND aggregates them into `NETWORK_PROPERTIES` for the
