+++ packages/protocol/src/testing/conformance/network/presence-same-state-no-double-fire.ts
@@ -1,7 +1,13 @@
 /**
  * Re-sending network/connect on an already-authenticated WS short-circuits
- * to buildHelloOk (`auth.handlers.ts:71`) which calls setOnline again.
- * The idempotency guard MUST suppress the redundant broadcast.
+ * to buildHelloOk in `identity/handlers/connect.handlers.ts`, which fires
+ * `presenceProjection.onAgentConnect(agentId, conn.id)` again. v10
+ * (architect plan #706) — the projection's redundant-connect rule
+ * (codex r3 P2 #1) handles this: when the existing entry's `connId`
+ * matches the incoming arg, the lifecycle predicate returns
+ * `prev === next`, the `emit` dedup gate elides, and no
+ * `presence/changed` is broadcast. This property asserts the
+ * idempotency guard holds.
  */
 import { Effect } from "effect";
 import { PROTOCOL_VERSION } from "../../../version.js";
@@ -27,7 +33,7 @@ export function registerSameStateNoDoubleFire(
     ctx,
     PRESENCE_CATEGORY,
     NAME,
-    "redundant setOnline (network/connect on already-authenticated WS) does NOT double-fire presence/changed",
+    "redundant onAgentConnect (network/connect on already-authenticated WS) does NOT double-fire presence/changed",
     sameStateNoDoubleFire(ctx, NAME),
   );
 }
